:truck:	chore: movendo função addProduct
